feat(theme): improvements and refactoring of theme module
fix: theme rebase
